+++ collections/composers/persistable.d.ts
@@ -1,0 +1,4 @@
+export function persistable({ beforeInsert, beforeUpdate }?: {
+    beforeInsert?: Function;
+    beforeUpdate?: Function;
+}): Function;